pyproject.toml
[project]
name = "devforge"
version = "0.1.0"
description = "Enterprise multi-agent engineering platform (capstone)"
requires-python = ">=3.12"

dependencies = [
    # Core framework
    "openai>=1.50.0",
    "openai-agents>=0.0.16",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    # Control plane (FastAPI + GitHub App)
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "mangum>=0.19.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=43.0.0",
    # MCP (Day 6)
    "mcp>=1.0.0",
]

[project.optional-dependencies]
# For local development — embeds + vectors run on your laptop.
local = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
]

# For deploy-to-AWS — no ML deps, just boto3.
aws = [
    "boto3>=1.40.0",
]

# All — use this during development so you can flip DEVFORGE_BACKEND freely.
all = [
    "devforge[local,aws]",
]
